backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.10.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
python-multipart>=0.0.6
pandas>=2.2.0
openpyxl>=3.1.2
